add gtest first smoke
add gtest simple EXPECT_EQ test
