+++ .travis.yml
@@ -36,11 +36,12 @@ script:
     - cd build
     - cmake .. -DCMAKE_BUILD_TYPE=$BUILD_CONFIGURATION -DTARGET_CPU=$TARGET_CPU
     - cmake --build .
+    - ctest -C Debug
+    - ctest -C Release
 
 
 after_script:
+    - uname --all
+    - env
     - pwd
-    - cd ~
-    - pwd
-    - cd build
     - ls -la
